+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## In progress
 
+- Vue asset sheet ([#150](https://github.com/ben/foundry-ironsworn/pull/150))
+
 ## 1.7.1
 
 - Vue progress/vow sheet ([#147](https://github.com/ben/foundry-ironsworn/pull/147))